requirements.txt
deepagents
langchain[openai]
akshare
python-dotenv
fastapi 
uvicorn
streamlit
apscheduler